backend/requirements.txt
django==4.1.6
djangorestframework==3.14.0
gunicorn==20.1.0
pysqlite==3.40.1